requirements.txt
customtkinter
reportlab
google-generativeai
python-dotenv
Pillow
pyinstaller